requirements.txt
python-binance==1.0.27
bt==1.1.0
beautifulsoup4==4.12.3
numpy==2.2.0
pandas==2.2.3
yfinance==0.2.51
matplotlib==3.10.0
hurst==0.0.5
kneed==0.8.5
plotly==5.24.1
seaborn==0.13.2
coinmetrics-api-client==2022.7.14.6
statsmodels==0.14.4
PyYAML==6.0.2
pyarrow==18.1.0
requests==2.32.3
